Criação dialogo de confirmação para excluir curso

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, Observable, of } from 'rxjs';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
 import { Course } from '../../model/course';
@@ -54,17 +55,26 @@ export class CoursesComponent implements OnInit {
   }
 
   onRemove(course: Course) {
-    this.service.remove(course._id).subscribe(
-      () => {
-        this._snackBar.open("Curso removido com sucesso!", 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: "Tem certeza que deseja remover esse curso ? ",
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if(result) {
+        this.service.remove(course._id).subscribe({
+          next: () => {
+            this._snackBar.open("Curso removido com sucesso!", 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+            this.refresh();
+          },
+          error: () => this.onError('Erro ao tentar remover curso!')
         });
-        this.refresh();
-      },
-      () => this.onError('Erro ao tentar remover curso!')
-    );
+      }
+    });
   }
 
 }
